requirements.txt
keyboard==0.13.5
pyperclip==1.8.2
pyautogui==0.9.54
requests==2.33.1
pystray==0.19.5
pillow==10.4.0
mss==10.1.0
paddlepaddle==2.6.2
paddleocr==2.10.0
opencv-python==4.6.0.66
numpy==2.4.4
pygetwindow==0.0.9
pytweening==1.2.0

# 测试依赖
pytest>=7.0.0
